+++ Apps/Contoso.WPF.Puppet.DotNetCore/App.xaml.cs
@@ -45,6 +45,7 @@ namespace Contoso.WPF.Puppet.DotNetCore
             }
 
             // Start AppCenter.
+            AppCenter.IsNetworkRequestsAllowed = Settings.Default.IsNetworkRequestsAllowed;
             AppCenter.Start("d967daf9-28ed-4899-84e8-17a00c064987", typeof(Analytics), typeof(Crashes));
             var userId = Settings.Default.UserId;
             if (!string.IsNullOrEmpty(userId))