+++ Apps/Contoso.WinForms.Puppet.DotNetCore/MainForm.cs
@@ -30,11 +30,13 @@ namespace Contoso.WinForms.Puppet.DotNetCore
             {
                 StorageMaxSizeTextBox.Text = Settings.Default.StorageMaxSize.ToString();
             }
+            AppCenterAllowNetworkRequests.Checked = Settings.Default.IsNetworkRequestsAllowed;
         }
 
         private void UpdateState()
         {
             AppCenterEnabled.Checked = AppCenter.IsEnabledAsync().Result;
+            AppCenterAllowNetworkRequests.Checked = AppCenter.IsNetworkRequestsAllowed;
             AnalyticsEnabled.Checked = Analytics.IsEnabledAsync().Result;
             CrashesEnabled.Checked = Crashes.IsEnabledAsync().Result;
             AnalyticsEnabled.Enabled = AppCenterEnabled.Checked;
@@ -44,6 +46,13 @@ namespace Contoso.WinForms.Puppet.DotNetCore
         private void AppCenterEnabled_CheckedChanged(object sender, EventArgs e)
         {
             AppCenter.SetEnabledAsync(AppCenterEnabled.Checked).Wait();
+        }
+
+        private void AppCenterAllowNetworkRequest_CheckedChanged(object sender, EventArgs e)
+        {
+            AppCenter.IsNetworkRequestsAllowed = AppCenterAllowNetworkRequests.Checked;
+            Settings.Default.IsNetworkRequestsAllowed = AppCenterAllowNetworkRequests.Checked;
+            Settings.Default.Save();
         }
 
         private void AnalyticsEnabled_CheckedChanged(object sender, EventArgs e)