+++ Apps/Contoso.WinForms.Puppet.DotNetCore/Program.cs
@@ -29,6 +29,7 @@ namespace Contoso.WinForms.Puppet.DotNetCore
             {
                 AppCenter.SetMaxStorageSizeAsync(storageMaxSize);
             }
+            AppCenter.IsNetworkRequestsAllowed = Settings.Default.IsNetworkRequestsAllowed;
             AppCenter.Start("7136db69-7f8d-4a14-90bd-12c9588ae0b9", typeof(Analytics), typeof(Crashes));
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);