+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Channel/Channel.cs
@@ -46,6 +46,8 @@ namespace Microsoft.AppCenter.Channel
 
         private bool _batchScheduled;
 
+        private bool _networkRequestsAllowed = true;
+
         private TimeSpan _batchTimeInterval;
 
         private readonly StatefulMutex _mutex = new StatefulMutex();
@@ -144,6 +146,27 @@ namespace Microsoft.AppCenter.Channel
             else
             {
                 Suspend(state, true, new CancellationException());
+            }
+        }
+
+        /// <summary>
+        /// Allow or disallow network requests.
+        /// </summary>
+        public bool IsNetworkRequestsAllowed
+        {
+            get => _networkRequestsAllowed;
+            set
+            {
+                _networkRequestsAllowed = value;
+                State state;
+                using (_mutex.GetLock())
+                {
+                    state = _mutex.State;
+                }
+                if (value && _enabled)
+                {
+                    CheckPendingLogs(state);
+                }
             }
         }
 
@@ -562,7 +585,11 @@ namespace Microsoft.AppCenter.Channel
                 AppCenterLog.Info(AppCenterLog.LogTag, "The service has been disabled. Stop processing logs.");
                 return;
             }
-
+            if (!_networkRequestsAllowed)
+            {
+                AppCenterLog.Debug(AppCenterLog.LogTag, "SDK is in offline mode.");
+                return;
+            }
             AppCenterLog.Debug(AppCenterLog.LogTag, $"CheckPendingLogs({Name}) pending log count: {_pendingLogCount}");
             using (_mutex.GetLock())
             {