+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Channel/IChannel.cs
@@ -43,6 +43,11 @@ namespace Microsoft.AppCenter.Channel
         void SetEnabled(bool enabled);
 
         /// <summary>
+        /// Allow or disallow network requests.
+        /// </summary>
+        bool IsNetworkRequestsAllowed { get; set; }
+
+        /// <summary>
         /// Stop all calls in progress and deactivate this channel
         /// </summary>
         Task ShutdownAsync();